third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-list-style-shorthand-expected.txt

Tests that the list-style shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('list-style') is "circle outside url(\"dummy://test.png\")"
PASS computedStyle.getPropertyValue('list-style') is "disc outside none"
PASS computedStyle.getPropertyValue('list-style') is "circle outside none"
PASS computedStyle.getPropertyValue('list-style') is "none outside none"
PASS computedStyle.getPropertyValue('list-style') is "circle outside url(\"dummy://test.png\")"
PASS successfullyParsed is true

TEST COMPLETE
